+++ R/import-standalone-cli_call_env.R
@@ -33,7 +33,7 @@ set_cli_abort_call <- function(env = rlang::caller_env()) {
   if (getOption("cli_abort_call") |> is.null()) {
     options(cli_abort_call = env)
     set_call <- as.call(list(function() options(cli_abort_call = NULL)))
-    do.call(on.exit, list(expr = set_call, after = FALSE), envir = env)
+    do.call(on.exit, list(expr = set_call, add = TRUE, after = FALSE), envir = env)
   }
   invisible()
 }